Updating python, getting dependencies updated
Python3.6 reached of of life in December of 2021, and Python3.7 will reach end
of life in June 2023. This commit changes the default Python to be 3.8 (end
of life is Oct 2024) and updates enough of the dependencies to get the docs
target to build correctly under tox.

Updating travisci

Blacken 6 files, update black to recent version for compatability with python3.8

isort

Making isort and black play nice together (line length)

Blacken has a new line lenght format

Removing usedevelop, this is causing tox to install things weirdly

Version bumps, doc changes

Missing lib

Setting isort length back to 88 to cut down on number of changes

Removing isort section, it was causing a warning message during linting phase

Commenting out the interop tests

There is a bigger issue that doesn't seem directly related to the python3.8 upgrade.
The interop tests are important, but I think the issue is due to a library being
updated, and debugging/fixing this will take away from the main point of this PR.

Will address in a followup PR.